# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConditionKind": {
        "enum": [
          "application_cooldown",
          "outreach_cooldown"
        ],
        "title": "ConditionKind",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "resume_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleCondition": {
        "additionalProperties": false,
        "properties": {
          "days": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/ConditionKind"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "days"
        ],
        "title": "RuleCondition",
        "type": "object"
      },
      "RuleCreate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleCondition"
              },
              {
                "type": "null"
              }
            ]
          },
          "lane_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lane Id"
          },
          "rule_type": {
            "$ref": "#/components/schemas/RuleType"
          },
          "scope": {
            "$ref": "#/components/schemas/RuleScope"
          },
          "statement": {
            "maxLength": 1000,
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "statement",
          "rule_type"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
      "RulePatch": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleCondition"
              },
              {
                "type": "null"
              }
            ]
          },
          "rule_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleType"
              },
              {
                "type": "null"
              }
            ]
          },
          "statement": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement"
          }
        },
        "title": "RulePatch",
        "type": "object"
      },
      "RuleResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleCondition"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lane_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lane Id"
          },
          "rule_type": {
            "$ref": "#/components/schemas/RuleType"
          },
          "scope": {
            "$ref": "#/components/schemas/RuleScope"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "company_id",
          "lane_id",
          "statement",
          "rule_type",
          "condition",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "RuleResponse",
        "type": "object"
      },
      "RuleScope": {
        "enum": [
          "global",
          "company",
          "lane"
        ],
        "title": "RuleScope",
        "type": "string"
      },
      "RuleType": {
        "enum": [
          "constraint",
          "preference",
          "cooldown"
        ],
        "title": "RuleType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Career OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/strategy-rules": {
      "get": {
        "operationId": "list_rules_api_v1_strategy_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleScope"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  },
                  "title": "Response List Rules Api V1 Strategy Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rules",
        "tags": [
          "strategy-rules"
        ]
      },
      "post": {
        "operationId": "create_rule_api_v1_strategy_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Rule",
        "tags": [
          "strategy-rules"
        ]
      }
    },
    "/api/v1/strategy-rules/{rule_id}": {
      "delete": {
        "operationId": "delete_rule_api_v1_strategy_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Rule",
        "tags": [
          "strategy-rules"
        ]
      },
      "patch": {
        "operationId": "edit_rule_api_v1_strategy_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Edit Rule",
        "tags": [
          "strategy-rules"
        ]
      }
    },
...
  }
}